config: redis client lazy connects (#796)
This changes all go services to lazy connect to redis, this means they
can boot when redis isn't available.

There is a tiny performance hit that the first request to hit the
service isn't primed with a ready redis connection. But if a service is
high traffic and the service needs to keep multiple connections running
to redis, then this delay would happen under normal conditions for
plenty of requests anyway.

Alternatives could be to create a new method, and migration services we
want over to use it. Or we could still call the ping method to try prime
one connection, but ignore/log the error.